router setup and code refactor as router and swagger tags replace on the router
