+++ database.py
@@ -13,3 +13,11 @@ SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False)
 
 # create base
 Base = declarative_base()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
